requirements.txt
streamlit==1.33.0
llama-cpp-python==0.2.57
pandas==2.2.1
scikit-learn==1.4.2
sentence-transformers==2.6.1
numpy==1.26.4
python-dotenv==1.0.1